fix: keep the manual route compatible with Next page exports
Next App Router pages cannot expose arbitrary named exports, so the
manual route was failing during Vercel's type validation. Moving the
manual draft helpers into a sibling module preserves the page behavior
and tests while keeping the route file to the allowed page exports.

Constraint: App Router page modules must only expose fields accepted by Next's page contract
Rejected: Remove helper tests entirely | would hide the route-export regression instead of fixing it
Confidence: high
Scope-risk: narrow
Reversibility: clean
Directive: Keep reusable logic for app routes in sibling modules when it needs named exports
Tested: bun run build
Tested: COREPACK_ENABLE_AUTO_PIN=0 corepack pnpm@9.15.4 verify
Not-tested: Vercel deployment pipeline beyond the local production build
